pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "visualization"
version = "0.0.1"
description = "General Visualization Tool for visualizing molecules and quantum chemistry calculations"
authors = [
    {name = "Piotr Kowalski", email = "hemiku@gmail.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Chemistry",
    "Topic :: Scientific/Engineering :: Visualization",
]
keywords = ["quantum-chemistry", "molecular-visualization", "orbitals", "sapt", "visualization"]

# Core runtime dependencies
dependencies = [
    "numpy>=1.20.0",
    "scipy>=1.6.0",
    "pyvista>=0.38.0",  # Default visualization backend
]

[project.optional-dependencies]
# GPU acceleration support
gpu = [
    "cupy-cuda11x>=10.0",
]

# Mayavi backend (optional, requires conda for easy install)
# Enable via: VIZ_BACKEND=mayavi environment variable
mayavi = [
    "mayavi>=4.7.0",
]

# Development dependencies
dev = [
    # Testing
    "pytest>=6.2.0",
    "pytest-cov>=2.12.0",
    "pytest-xdist>=2.3.0",

    # Code quality
    "black>=21.6b0",
    "flake8>=3.9.0",
    "mypy>=0.910",

    # Jupyter support
    "ipykernel>=6.0.0",

    # Runtime deps (explicitly included for dev)
    "numpy>=1.20.0",
    "scipy>=1.6.0",
]

# All optional dependencies combined
all = [
    "cupy-cuda11x>=10.0",
    "mayavi>=4.7.0",
    "pytest>=6.2.0",
    "pytest-cov>=2.12.0",
    "pytest-xdist>=2.3.0",
    "black>=21.6b0",
    "flake8>=3.9.0",
    "mypy>=0.910",
    "ipykernel>=6.0.0",
]

[project.urls]
Homepage = "https://github.com/hemiku/Visualization/"
Repository = "https://github.com/hemiku/Visualization/"
Issues = "https://github.com/hemiku/Visualization/issues"

[tool.hatchling.build.targets.wheel]
packages = ["visualization"]

[tool.pytest.ini_options]
# Pytest configuration
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
testpaths = ["tests"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--color=yes",
]
markers = [
    "unit: Unit tests for individual components",
    "integration: Integration tests for workflows",
    "validation: Scientific correctness validation tests",
    "slow: Tests that take significant time to run",
    "gpu: Tests requiring GPU/CuPy",
    "requires_examples: Tests requiring example data files",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]
minversion = "3.7"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.coverage.run]
source = ["visualization"]
omit = [
    "*/tests/*",
    "*/test_*.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
